+++ driftbase/match_placements.py
@@ -368,7 +368,6 @@ def process_gamelift_queue_event(queue_name: str, message: dict):
 
 def process_match_message(queue_name: str, message: dict):
     log.debug(f"match-placements::process_match_message() received event in queue '{queue_name}': '{message}'")
-
     event = message["event"]
     if event == "match_status_changed":
         match_id = message.get("match_id", None)
@@ -383,8 +382,6 @@ def process_match_message(queue_name: str, message: dict):
 
         if match_status == "ended":
             return _process_match_ended(match_id)
-    else:
-        log.error(f"Unexpected event '{event}' published.")
 
 # Helpers
 
@@ -565,7 +562,7 @@ def _process_fulfilled_queue_event(event_details: dict):
         elif party_id:
             log.info(f"Placement '{placement_id}' completed for party '{party_id}'")
 
-            party_member_ids = placement["party_member_ids"]
+            party_member_ids = placement["player_ids"]
 
             for party_member_id in party_member_ids:
                 if party_member_id not in connection_options_by_player_id:
@@ -634,7 +631,7 @@ def _process_cancelled_queue_event(event_details: dict):
         elif party_id:
             log.info(f"Placement '{placement_id}' cancelled for party '{party_id}'")
 
-            party_member_ids = placement["party_member_ids"]
+            party_member_ids = placement["player_ids"]
             _post_match_placement_event_to_members(party_member_ids, "MatchPlacementCancelled", placement)
         else:
             log.info(f"Placement '{placement_id}' cancelled for player '{player_id}'")
@@ -681,7 +678,7 @@ def _process_timed_out_queue_event(event_details: dict):
         elif party_id:
             log.info(f"Placement '{placement_id}' timed out for party '{party_id}'")
 
-            party_member_ids = placement["party_member_ids"]
+            party_member_ids = placement["player_ids"]
             _post_match_placement_event_to_members(party_member_ids, "MatchPlacementTimedOut", placement)
         else:
             log.info(f"Placement '{placement_id}' timed out for player '{player_id}'")
@@ -727,7 +724,7 @@ def _process_failed_queue_event(event_details: dict):
         elif party_id:
             log.info(f"Placement '{placement_id}' failed for party '{party_id}'")
 
-            party_member_ids = placement["party_member_ids"]
+            party_member_ids = placement["player_ids"]
             _post_match_placement_event_to_members(party_member_ids, "MatchPlacementFailed", placement)
         else:
             log.info(f"Placement '{placement_id}' failed for player '{player_id}'")